chore(ColorRamp): rename colorramp
