+++ src/ColorRamp.ts
